+++ sctplib/programs/echo_monitor.c
@@ -1122,7 +1122,7 @@ static void resize(int sig)
 
 int main(int argc, char **argv)
 {
-  short sctpInstance;
+  int sctpInstance;
   SCTP_ulpCallbacks echoUlp;
   SCTP_InstanceParameters instanceParameters;
   unsigned int index;
